refactor: Sorry, big dump
